Move tests for `near-vm-logic` from `/tests` to `/src/tests/`
