+++ runtime/near-vm-logic/Cargo.toml
@@ -40,32 +40,3 @@ protocol_feature_alt_bn128 = ["bn", "near-primitives-core/protocol_feature_alt_b
 
 # Use this feature to enable counting of fees and costs applied.
 costs_counting = []
-
-[[test]]
-name = "test_storage_read_write"
-path = "tests/test_storage_read_write.rs"
-
-[[test]]
-name = "test_context"
-path = "tests/test_context.rs"
-
-[[test]]
-name = "test_miscs"
-path = "tests/test_miscs.rs"
-required-features = ["costs_counting"]
-
-[[test]]
-name = "test_registers"
-path = "tests/test_registers.rs"
-
-[[test]]
-name = "test_storage_usage"
-path = "tests/test_storage_usage.rs"
-
-[[test]]
-name = "test_promises"
-path = "tests/test_promises.rs"
-
-[[test]]
-name = "test_iterators"
-path = "tests/test_iterators.rs"
