+++ runtime/near-vm-logic/src/lib.rs
@@ -8,6 +8,8 @@ pub mod gas_counter;
 mod logic;
 pub mod mocks;
 pub mod serde_with;
+#[cfg(test)]
+mod tests;
 pub mod types;
 mod utils;
