+++ runtime/near-vm-logic/src/tests/iterators.rs
@@ -1,9 +1,6 @@
-mod fixtures;
-mod vm_logic_builder;
-
-use crate::fixtures::get_context;
+use crate::tests::fixtures::get_context;
+use crate::tests::vm_logic_builder::VMLogicBuilder;
 use near_vm_errors::{HostError, VMLogicError};
-use vm_logic_builder::VMLogicBuilder;
 
 #[test]
 fn test_iterator_deprecated() {
