+++ runtime/near-vm-logic/src/tests/mod.rs
@@ -1,0 +1,12 @@
+mod context;
+mod fixtures;
+mod gas_counter;
+mod helpers;
+mod iterators;
+mod miscs;
+mod promises;
+mod registers;
+mod storage_read_write;
+mod storage_usage;
+mod view_method;
+mod vm_logic_builder;
