+++ runtime/near-vm-logic/src/tests/promises.rs
@@ -1,12 +1,8 @@
-mod fixtures;
-mod helpers;
-mod vm_logic_builder;
-
-use fixtures::get_context;
-use helpers::*;
-use near_vm_logic::types::PromiseResult;
+use crate::tests::fixtures::get_context;
+use crate::tests::helpers::*;
+use crate::tests::vm_logic_builder::VMLogicBuilder;
+use crate::types::PromiseResult;
 use serde_json;
-use vm_logic_builder::VMLogicBuilder;
 
 #[test]
 fn test_promise_results() {
