+++ runtime/near-vm-logic/src/tests/registers.rs
@@ -1,10 +1,7 @@
-mod fixtures;
-mod vm_logic_builder;
-
-use fixtures::get_context;
+use crate::tests::fixtures::get_context;
+use crate::tests::vm_logic_builder::VMLogicBuilder;
+use crate::VMConfig;
 use near_vm_errors::{HostError, VMLogicError};
-use near_vm_logic::VMConfig;
-use vm_logic_builder::VMLogicBuilder;
 
 #[test]
 fn test_one_register() {
