+++ runtime/near-vm-logic/src/tests/view_method.rs
@@ -1,8 +1,5 @@
-mod fixtures;
-mod vm_logic_builder;
-
-use crate::fixtures::get_context;
-use vm_logic_builder::VMLogicBuilder;
+use crate::tests::fixtures::get_context;
+use crate::tests::vm_logic_builder::VMLogicBuilder;
 
 macro_rules! test_prohibited {
     ($f: ident $(, $arg: expr )* ) => {
